+++ lib/src/modbus/modbus.g.dart
@@ -47,3 +47,28 @@ Map<String, dynamic> _$$ModbusParameterImplToJson(
       'dataAddress': instance.dataAddress,
       'dataLength': instance.dataLength,
     };
+
+_$ModbusParameterInputImpl _$$ModbusParameterInputImplFromJson(
+        Map<String, dynamic> json) =>
+    _$ModbusParameterInputImpl(
+      schema: json['schema'] == null
+          ? ModbusParameterSchema.single
+          : const ModbusParameterSchemaConverter()
+              .fromJson(json['schema'] as String),
+      splitEach: json['splitEach'] as String?,
+      controllerAddress: json['controllerAddress'] as String? ?? '',
+      functionCode: json['functionCode'] as String? ?? '',
+      dataAddress: json['dataAddress'] as String? ?? '',
+      dataLength: json['dataLength'] as String? ?? '',
+    );
+
+Map<String, dynamic> _$$ModbusParameterInputImplToJson(
+        _$ModbusParameterInputImpl instance) =>
+    <String, dynamic>{
+      'schema': const ModbusParameterSchemaConverter().toJson(instance.schema),
+      'splitEach': instance.splitEach,
+      'controllerAddress': instance.controllerAddress,
+      'functionCode': instance.functionCode,
+      'dataAddress': instance.dataAddress,
+      'dataLength': instance.dataLength,
+    };